add routing in the app
