+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Microverse\\React\\math-magicians\\src\\index.js":"1","D:\\Microverse\\React\\math-magicians\\src\\reportWebVitals.js":"2","D:\\Microverse\\React\\math-magicians\\src\\logic\\calculate.js":"3","D:\\Microverse\\React\\math-magicians\\src\\logic\\operate.js":"4"},{"size":512,"mtime":1612805550106,"results":"5","hashOfConfig":"6"},{"size":373,"mtime":1612287175826,"results":"7","hashOfConfig":"6"},{"size":1052,"mtime":1612989769059,"results":"8","hashOfConfig":"6"},{"size":591,"mtime":1612989769059,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"15l87cj",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Microverse\\React\\math-magicians\\src\\index.js",[],"D:\\Microverse\\React\\math-magicians\\src\\reportWebVitals.js",[],"D:\\Microverse\\React\\math-magicians\\src\\logic\\calculate.js",[],"D:\\Microverse\\React\\math-magicians\\src\\logic\\operate.js",[]]
+[{"D:\\Microverse\\React\\math-magicians\\src\\index.js":"1","D:\\Microverse\\React\\math-magicians\\src\\reportWebVitals.js":"2","D:\\Microverse\\React\\math-magicians\\src\\logic\\calculate.js":"3","D:\\Microverse\\React\\math-magicians\\src\\logic\\operate.js":"4"},{"size":556,"mtime":1613006909895,"results":"5","hashOfConfig":"6"},{"size":373,"mtime":1612287175826,"results":"7","hashOfConfig":"6"},{"size":1052,"mtime":1612989769059,"results":"8","hashOfConfig":"6"},{"size":591,"mtime":1612989769059,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"15l87cj",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"D:\\Microverse\\React\\math-magicians\\src\\index.js",[],"D:\\Microverse\\React\\math-magicians\\src\\reportWebVitals.js",[],["19","20","21","22","23","24","25"],"D:\\Microverse\\React\\math-magicians\\src\\logic\\calculate.js",[],"D:\\Microverse\\React\\math-magicians\\src\\logic\\operate.js",[],{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},"jsx-a11y/accessible-emoji",[],"lines-around-directive",["40"],"no-spaced-func",["41"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
