+++ src/components/App.js
@@ -1,24 +1,18 @@
-import React, { useState } from 'react';
-import '../App.css';
-import ButtonPanel from './buttonPanel';
-import Display from './display';
-import calculate from '../logic/calculate';
+import { Switch, Route } from "react-router-dom";
+import Calculator from "./calculator";
+import Home from "./home";
+import Navbar from "./navbar";
+import Quote from "./quote";
 
 const App = () => {
-
-  const calculator = {total: '0', next: '0', operation: ''};
-  const [state, setState] = useState(calculator);
-
-  const handleClick = buttonName => {
-    console.log(state);
-    const data = calculate(state, buttonName);
-    setState({ total: data.total, next: data.next, operation: data.operation });
-  };
-
   return (
     <>
-     <Display result={state.total} />
-     <ButtonPanel clickHandler={handleClick}/>
+      <Navbar/>
+      <Switch>
+        <Route component = {Home} exact path="/"/>
+        <Route component = {Calculator} path="/calculator"/>
+        <Route component = {Quote} path="/quote"/>
+      </Switch>
     </>
   ) 
 }
