+++ src/components/home.js
@@ -1,0 +1,7 @@
+const Home = () => {
+  return ( 
+    <h3>Home</h3>
+   );
+}
+ 
+export default Home;
