+++ src/components/quote.js
@@ -1,0 +1,7 @@
+const Quote = () => {
+  return ( 
+    <div>Quote</div>
+   );
+}
+ 
+export default Quote;
