+++ src/styles/style.css
@@ -4,7 +4,7 @@
 
 .display {
 	background-color: gray;
-	width: 75%;
+	width: 700px;
 	margin: 5% auto 0;
 	padding: 15px;
 	text-align: right;
@@ -15,6 +15,6 @@
 	display: flex;
 	flex-direction: column;
 	align-items: center;
-	width: 75%;
+	width: 700px;
 	margin: 0 auto;
 }
